Expose raw Soketto interface via ws_client for more control where needed (eg in tests)
